refactor: Package TagList component
